% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFrac.R
\docType{data}
\name{COHORT}
\alias{COHORT}
\title{Example of cohort description input file}
\format{An object of class \code{data.frame} with 5 rows and 7 columns.}
\usage{
COHORT
}
\description{
Example of cohort description input file
}
\author{
Vincent Laville \email{vincent.laville@pasteur.fr}
}
\keyword{data}
