% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{getAdditivelyCodedMatrix}
\alias{getAdditivelyCodedMatrix}
\title{Code additively the genotypes
Sum the two columns per individual of the input matrix (each corresponding to one chromosome)
to get the additive coding of the individual at the locus.}
\usage{
getAdditivelyCodedMatrix(genoMat, lind)
}
\arguments{
\item{genoMat}{is the matrix of genotype with two colums per individual
Rows are variants and columns are individuals}

\item{lind}{is the list of individual in the genotype matrix}
}
\value{
The additively coded genotyped matrix
}
\description{
Code additively the genotypes
Sum the two columns per individual of the input matrix (each corresponding to one chromosome)
to get the additive coding of the individual at the locus.
}
\examples{
genoMat = matrix(rbinom(5*10,1,runif(10,0,1)),nrow=5,ncol=10)
lind = c("ind1","ind2","ind3","ind4","ind5")
colnames(genoMat) = rep(lind,each=2)
addCodedMat = getAdditivelyCodedMatrix(genoMat, lind)

}
