% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{getMatCorSVD}
\alias{getMatCorSVD}
\title{Perform singular Value Decomposition on the correlation matrix}
\usage{
getMatCorSVD(cormat, k = qr(cormat)$rank)
}
\arguments{
\item{cormat}{is the correlation matrix}

\item{k}{is the number of eigenvectors to keep. 
Default is the correlation matrix rank.}
}
\value{
A list with 
\describe{
  \item{eigval}{A vector of the top \code{k} eigenvalues}
  \item{eigvev}{A matrix of the top \code{k} eigenvectors}
}
}
\description{
Perform singular Value Decomposition on the correlation matrix
}
\examples{
a <- matrix(runif(10*10, 1, 5), nrow = 10)
matcor <- cor(a)
getMatCorSVD(matcor)
getMatCorSVD(matcor, k = 4)

}
