% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFrac.R
\docType{data}
\name{ind}
\alias{ind}
\title{This is data included in the package}
\description{
This is data included in the package
}
\references{
\url{https://github.com/slowkow/proxysnps/blob/master/data/ind.rda}
}
\author{
Kamil Slowikowski \email{kslowikowski@fas.harvard.edu}
}
\keyword{data}
