% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varFrac.R
\name{standardizeBeta}
\alias{standardizeBeta}
\title{Derive betas in the standardized model from betas in the general model}
\usage{
standardizeBeta(betaG, betaINT, maf, meanE, varE, type)
}
\arguments{
\item{betaG}{is the vector of main genetic effects in the general model}

\item{betaINT}{is the vector of interaction effects in the general model}

\item{maf}{is the vector of the variants' frequency}

\item{meanE}{is the mean of the exposure}

\item{varE}{is the variance of the exposure}

\item{type}{designates the coefficients to standardize:
"G" for the main genetic effect and "I" for the interaction effects}
}
\value{
The vector of standardized effects
}
\description{
Derive betas in the standardized model from betas in the general model
}
\examples{
betaGs <- rnorm(10, 0, 0.1)
betaIs <- rnorm(10, 0, 0.05)
mafs <- runif(10, 0.05, 0.95)
meanE <- runif(1, -2, 2)
varE <- runif(1, 0.5, 1.5)
std_betaG <- standardizeBeta(betaGs, betaIs, mafs, meanE, varE, "G")
std_betaI <- standardizeBeta(betaGs, betaIs, mafs, meanE, varE, "I")

tryCatch(
  std_betaI <- standardizeBeta(betaGs, betaIs, mafs, meanE, varE, 
                               "anything different from G or I"),
  error = function(e) print(e))

}
